pyproject.toml
[project]
name = "soc_analyst"
version = "1.0.0"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0.0",
]
